Add linting (#70)
Adds ESLint, based on the Nava ESLint base, configured to run against TypeScript.

Plugins:
    React
    Jest & Jest-dom
    Testing Library
    Typescript & Typscript Type checking
    Prettier

Using the recommended settings for all of the above & applies the rules to the existing repo.

The only custom rule is "react/react-in-jsx-scope": "off", which is recommended for our NextJS setup. We can add others as we discover them.

This PR also converts the placeholder app files to tsx and turns on strict mode.